Compiling with `g++` force explicit cast. Why not... (#127)
* Fix filter

* Calloc cast

* Cast btx_message_iterator_t

* Fix crosses initialization

* common data

* calloc

* utarray cast

* Fix remaining

---------

+++ test/state_machine/cases_push_messages/1.source_callbacks.c
@@ -2,7 +2,7 @@
 
 struct data_s {
   uint64_t i;
-  int count;
+  uint64_t count;
 };
 
 typedef struct data_s data_t;
